generalized bent cigar and discus functions are now defined in their own f_..._generalized.c files
