+++ code-experiments/src/suite_largescale.c
@@ -6,8 +6,8 @@
 #include "coco.h"
 
 #include "f_ellipsoid.c"
-#include "f_discus.c"
-#include "f_bent_cigar.c"
+#include "f_discus_generalized.c"
+#include "f_bent_cigar_generalized.c"
 #include "f_different_powers.c"
 
 static coco_suite_t *coco_suite_allocate(const char *suite_name,
@@ -23,7 +23,7 @@ static coco_suite_t *suite_largescale_initialize(void) {
   
   coco_suite_t *suite;
   const size_t dimensions[] = { 40, 80, 160, 320, 640, 1280, 2560, 5120};
-  suite = coco_suite_allocate("bbob-largescale", 4, 8, dimensions, "instances:1-15");
+  suite = coco_suite_allocate("bbob-largescale", 4, 8, dimensions, "instances:1-5");
   return suite;
 }
